Added high_accuracy metric
