--- config.ini
@@ -1,11 +1,0 @@
-[tweepy]
-consumer_key = 
-consumer_secret = 
-access_token = 
-access_secret = 
-
-[software]
-locations_path = data\locations.sql
-database_path = data\data.sql
-shp_path = shapefiles\boundary_ncr.shp
-
